requirements.txt
Babel==2.4.0
backports-abc==0.5
bleach==2.0.0
decorator==4.1.2
Django==1.11.3
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.6.1
django-filter==1.0.4
django-floppyforms==1.7.0
django-jquery==3.1.0
django-material==1.0.0
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
entrypoints==0.2.3
html5lib==0.999999999
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.2
pexpect==4.2.1
phonenumberslite==8.6.0
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
selenium==3.5.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.0
xlwt==1.2.0
